Python Password Power
Project Overview: This project is a testament to my efforts in integrating secure password principles with my proficiency in Python. The objective was to create an intricate password strength verifier that not only validates the robustness of the input password but also cross-verifies it against a database of known compromised passwords. The primary goal was to encourage the use of robust and secure passwords by providing real-time feedback to the user about the potential weaknesses in their chosen password.

Throughout the process of developing this tool, I incorporated an array of security features that a strong password should possess: a minimum length requirement, the presence of both uppercase and lowercase characters, the inclusion of numerals, special characters, and constraints against using repetitive or consecutive characters.

By achieving this, the project stands as an intersection of security principles and Python programming, thus ensuring that a user's chosen password can withstand a variety of common attacks while also promoting safer password practices.